Add files via upload
On the Module 2 Challenge in UCI Bootcamp spot, the example showed no conditional formatting for the percent change. However, the rubric mentioned that 10 points are added for that. So I added it.

There are bas and txt files for each module. Module 5 is the VBA script that contains all of the code.

I included a DIY for fun: m2_multiple_year_stock_data_diy.xlsm.
This one is my real worksheet: m2_multiple_year_stock_data.xlsm (without the diy at the end of the file name).